+++ tests/testthat/test-property-stdev.R
@@ -1,14 +1,14 @@
 
 test_that("Network is invoked in priority: stdev", {
   # First look for `variance`, then invoke algorithm if not found.
-  d <- distribution(
+  d <- suppressWarnings(distribution(
     density = function(x) {
       stats::dnorm(x, sd = 3)
     },
     range = c(-Inf, Inf),
     variance = 100, # deliberately incorrect
     .vtype = "continuous"
-  )
+  ))
   expect_equal(stdev(d), 10)
   expect_equal(eval_stdev_from_network(d), 10)
   d$variance <- NULL